Create code for other resource endpoints 3

+++ bookms/src/main/java/com/edu/bookms/api/exception/BookNotFoundException.java
@@ -1,7 +1,7 @@
 package com.edu.bookms.api.exception;
 
 public class BookNotFoundException extends RuntimeException {
-    public BookNotFoundException(Integer id) {
+    public BookNotFoundException(Long id) {
         super("Could not find book " + id);
     }
